requirements.txt
datasets==3.3.1
openai==0.28.0
python-dotenv==1.0.1
